refactor: in/out
