+++ dispatcher/src/main/java/com/github/jon7even/service/in/AuthorizationService.java
@@ -1,4 +1,4 @@
-package com.github.jon7even.service;
+package com.github.jon7even.service.in;
 
 import com.github.jon7even.dto.UserShortDto;
