+++ dispatcher/src/main/java/com/github/jon7even/service/in/UpdateProducerService.java
@@ -1,4 +1,4 @@
-package com.github.jon7even.service;
+package com.github.jon7even.service.in;
 
 import org.telegram.telegrambots.meta.api.objects.Update;
