+++ dispatcher/src/test/java/jon7even/service/AuthorizationServiceTest.java
@@ -7,8 +7,8 @@ import com.github.jon7even.dto.UserShortDto;
 import com.github.jon7even.mapper.UserMapper;
 import com.github.jon7even.model.user.UserEntity;
 import com.github.jon7even.repository.UserRepository;
-import com.github.jon7even.service.AuthorizationService;
-import com.github.jon7even.service.AuthorizationServiceImpl;
+import com.github.jon7even.service.in.AuthorizationService;
+import com.github.jon7even.service.in.AuthorizationServiceImpl;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
