+++ node/src/main/java/com/github/jon7even/service/MainServiceImpl.java
@@ -4,8 +4,8 @@ import com.github.jon7even.cache.UserDataCache;
 import com.github.jon7even.model.company.CompanyEntity;
 import com.github.jon7even.repository.CompanyRepository;
 import com.github.jon7even.repository.UserRepository;
-import com.github.jon7even.service.message.ReplyMessageService;
-import com.github.jon7even.service.producer.SenderMessageService;
+import com.github.jon7even.service.in.message.ReplyMessageService;
+import com.github.jon7even.service.out.producer.SenderMessageService;
 import com.github.jon7even.telegram.BotState;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
